pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "runtime-contracts"
version = "0.4.0"
description = "Declarative runtime contract validator with promotion-readiness scoring"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Proto Labs" }]
keywords = ["runtime", "contracts", "mcp", "healthcheck", "validation", "dns", "tls", "disk"]
dependencies = [
  "anyio>=4.0",
  "httpx>=0.27",
  "mcp>=1.0",
]

[project.optional-dependencies]
config = ["pyyaml>=6.0"]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "pyyaml>=6.0"]

[project.scripts]
contract-check = "runtime_contracts.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.coverage.run]
source = ["runtime_contracts"]
branch = true
